ROGO-3228 Previous attempt information should not appear on Summative exams
Before the refactor of the user_index page to use templates the code did not output anything related to previous attempts on the page for summative exams. It looks as though we missed that this was part of the logic during the refactor.

After this change the page should display information in the same way as it did in Rogo 7.3.0

(cherry picked from commit aaa151985d8af2b995d354daaa943d94c039b1d3)