EXAMSYS-3269 Improve reporting
The code should now be easier to follow.

Bugs have been fixed where it would not find images if:

* There was no hieght
* There was no witdh
* There was no alt text
* The image was not from a rogo_directory

It will also check other locations in ExamSys if the image is stored in a directory other than the help directory.

It will also report images that are not in a rogo_directory as such.

(cherry picked from commit 0ab46741e3a27b285951050b75a2eedbe03af726)