EXAMSYS-3492 Normalise report start time results
Many reports selected the paper attempt starttimes in different ways.

For example:

* Only including attempts that started one or after the start time
* Giving a two minute buffer before the start time for including attempts
* Giving a two minute buffer for summative papers only

Summative exams and progress tests can both be started by invigilators who may have a clock that is not the same as the system time for ExamSys, this means that they can be started early.

We are normalising the reports so that they will give the two minute grace period to summative and progress papers, but not to other types.

We are also making it easier to adjust which types of paper have which grace period in the future.