EXAMSYS-3603 Frequency and discrimination report should not break for questions with no parts
Before this change if a fill in the blank question had no blanks defined it would break the report because the code used to display the exclusion options would throw a type exception, because it very often assumes that parts exist.

This chane ensures that all questions will always have a parts array, even if it is empty.